requirements/base.txt
Django~=4.0.3
dj-database-url~=0.5.0
django-localflavor~=3.1
django-phonenumber-field[phonenumberslite]~=6.1.0
django-simple-history~=3.0.0
python-decouple~=3.6
psycopg2~=2.9.3
libsass~=0.21
django-compressor~=3.1 
django-sass-processor~=1.1
